# Optimize zlib reader and consolidate sync.pools