--- a/plumbing/format/objfile/writer.go
+++ b/plumbing/format/objfile/writer.go
@@ -5,9 +5,9 @@ import (
 	"errors"
 	"io"
 	"strconv"
-	"sync"
 
 	"github.com/go-git/go-git/v5/plumbing"
+	"github.com/go-git/go-git/v5/utils/sync"
 )
 
 var (
@@ -21,7 +21,7 @@ type Writer struct {
 	raw    io.Writer
 	hasher plumbing.Hasher
 	multi  io.Writer
-	zlib   io.WriteCloser
+	zlib   *zlib.Writer
 
 	closed  bool
 	pending int64 // number of unwritten bytes
@@ -32,19 +32,11 @@ type Writer struct {
 // The returned Writer implements io.WriteCloser. Close should be called when
 // finished with the Writer. Close will not close the underlying io.Writer.
 func NewWriter(w io.Writer) *Writer {
-	zlib := zlibPool.Get().(*zlib.Writer)
-	zlib.Reset(w)
-
+	zlib := sync.GetZlibWriter(w)
 	return &Writer{
 		raw:  w,
 		zlib: zlib,
 	}
-}
-
-var zlibPool = sync.Pool{
-	New: func() interface{} {
-		return zlib.NewWriter(nil)
-	},
 }
 
 // WriteHeader writes the type and the size and prepares to accept the object's
@@ -110,7 +102,7 @@ func (w *Writer) Hash() plumbing.Hash {
 // Calling Close does not close the wrapped io.Writer originally passed to
 // NewWriter.
 func (w *Writer) Close() error {
-	defer zlibPool.Put(w.zlib)
+	defer sync.PutZlibWriter(w.zlib)
 	if err := w.zlib.Close(); err != nil {
 		return err
 	}